perf(frontend): frame-split the agent chat panel; region skeletons + fixes
Split AgentChatPanel into a synchronous frame (Splitter + Tabs + region
slots) and a lazy AgentConversation body, so the real structure paints
immediately and only the heavy AI-SDK/useChat chunk loads behind a
transcript/composer skeleton. Session bar and rail stay lazy inside the
frame, each behind its own region skeleton, and every region eases in via
a new opacity-only MountFade instead of a hard Suspense pop. Removes the
crossfade AgentChatPanelHost; the composed AgentChatSkeleton now serves
only the pre-confirmation gate in MainLayout.

Also fixes three skeleton-state defects:
- reserve the config-header kebab's footprint (dynamic loading fallback)
  so it no longer pops in and shifts Commit/Deploy.
- trim the Model & harness skeleton row so it stops overflowing the 400px
  config panel.
- restore scroll-to-tab: instant scroll now applies only to the bar's
  initial reload restore, so a newly-created tab glides into view again.
